migrate test for openapi compatibility
